Allow jenkins to run any command in sudo
This is needed to be able to setup dynamic nodes and inject system wide
variables and files. Main need for this is to add a CA certificate file
for Docker registry.

Makes things less secure, but since I am mainly using the Docker cloud
provider, the container is short lived